Added database code for voting in progress

+++ database_support/database_helper.rb
@@ -3,9 +3,12 @@ require 'capybara_support/configuration'
 
 def connection(options)
   query_name = options[:query_name] || :SDP
+  @database_name = options[:database_name] || :comments
 
   get_environment()
   @port_number = 5432
+  puts "Database name is: #{@database_name}"
+  puts "Query name is: #{query_name}"
   @conn = PGconn.connect(@ip_address, @port_number, '', '', @database_name, @user_name, @password)
 
   sql = query_collection(query_name)
@@ -19,25 +22,63 @@ def connection(options)
   return output
 end
 
+#private
+#def get_environment1()
+#  env = $environment #Global variable from configuration file
+#  case env
+#    when :stage
+#      @ip_address = ""
+#      @database_name = ""
+#      @user_name = ""
+#      @password = ""
+#
+#    when :production
+#      @ip_address = ""
+#      @database_name = ""
+#      @user_name = ""
+#      @password = ""
+#
+#    when :preview
+#      @ip_address = ""
+#      @database_name = ""
+#      @user_name = ""
+#      @password = ""
+#
+#    when :demo
+#      @ip_address = "192.168.113.22"
+#      @database_name = "comments"
+#      @user_name = "comments"
+#      @password = "alfjljqreovcab408qewfasdlj"
+#  end
+#  @url = CapybaraSupport::Configuration.get_environment_url
+#end
+
 private
 def get_environment()
   env = $environment #Global variable from configuration file
+  if (@database_name == :comments)
+    comments_database(env)
+  else
+    BTB_database(env)
+  end
+  @url = CapybaraSupport::Configuration.get_environment_url
+end
+
+private
+def comments_database(env)
   case env
     when :stage
       @ip_address = ""
-      @database_name = ""
       @user_name = ""
       @password = ""
 
     when :production
       @ip_address = ""
-      @database_name = ""
       @user_name = ""
       @password = ""
 
     when :preview
       @ip_address = ""
-      @database_name = ""
       @user_name = ""
       @password = ""
 
@@ -47,20 +88,63 @@ def get_environment()
       @user_name = "comments"
       @password = "alfjljqreovcab408qewfasdlj"
   end
-  @url = CapybaraSupport::Configuration.get_environment_url
+end
+
+private
+def BTB_database(env)
+  case env
+    when :stage
+      @ip_address = ""
+      @user_name = ""
+      @password = ""
+
+    when :production
+      @ip_address = ""
+      @user_name = ""
+      @password = ""
+
+    when :preview
+      @ip_address = ""
+      @user_name = ""
+      @password = ""
+
+    when :demo
+      @ip_address = "btb.demo.modcloth.com"
+      @database_name = "btb"
+      @user_name = "btb"
+      @password = "mcuemcvhbuoaec"
+  end
 end
 
 private
 def query_collection(query_name)
   case query_name
     when :SDP
-     sql = "SELECT commentable_id, commentable_name, id
+      sql = "SELECT commentable_id, commentable_name, id
      FROM comments where account_email !=" + "'" + $email + "'" + "
      AND status = 'active'
      AND agreement_count = 0 order by id desc limit 1;"
 
     when :DeleteComments
-    sql = "delete from comments where account_email =" + "'" + $email + "'" + ""
+      sql = "delete from comments where account_email =" + "'" + $email + "'" + ""
+
+    when :Voting_in_progress_SampleDetails
+      sql = "SELECT s.name, s.price, s.voting_starts_at,date_trunc('hour',s.voting_ends_at - now()), count(v.product_id)
+             FROM samples s FULL OUTER JOIN votes v ON s.product_id = v.product_id
+             WHERE state ='active' AND (voting_starts_at <= now() AND voting_ends_at > now())
+             GROUP BY v.product_id, s.name, s.voting_starts_at, s.price, s.voting_starts_at,s.voting_ends_at
+             ORDER BY s.voting_starts_at DESC
+             LIMIT 1"
+
+    when :Voting_in_progress_SampleCount
+      sql = "SELECT count(*)
+             FROM samples
+             WHERE state ='active' AND (voting_starts_at <= now() AND voting_ends_at > now())"
+
+    when :Voting_in_progress_CommentCount
+      commentable_name = "Sample 2031".gsub("Sample ",'')  #Need to include this line in voting in progress scripts
+      puts "Commentable name is:#{commentable_name}"
+      sql = "SELECT count(*) FROM comments where commentable_name = " + "'" + commentable_name + "'" + ""
 
     else
       print "\n No matching query..Please check your typos.... \n"
@@ -82,8 +166,19 @@ def query_result(query_name, res)
     when :DeleteComments
       puts "Comments deleted successfully for user #{$email}"
 
+    when :Voting_in_progress_SampleDetails
+      $sample_name = res.getvalue(0, 0)
+      $sample_price = res.getvalue(0, 1)
+      $voting_time =  res.getvalue(0, 2)
+      $vote_count = res.getvalue(0, 3)
+      puts "Sample name is ****************- #{$sample_name}"
+      puts "Sample price is ****************- #{$sample_price}"
+      puts "Vote count is ****************- #{$vote_count}"
+      puts "voting time is ****************- #{$voting_time}"
+
     else
       value = res.getvalue(0, 0)
       return value
   end
 end
+
